Add data loading (WIP)
Attempt to load the dt-helix.raw data (parsed and exported as a binary
file, which is included as bytes and deserialized in the file). Two
"helixes" are included: out1.bincode and out2.bincode. Because the Dxx,
Dyy and Dzz components seem to be messing up everything, the particles
just end up floating away in that one.

Problem: The data seems to be separated on the x and y axes, which means
that the particles don't form a helix but instead some disjointed
shapes. Will have to fix that.